Turn BookPage and MakerPage into modules
