+++ frontend/src/components/BookPage/index.ts
@@ -1,0 +1,4 @@
+import BookPage from './BookPage';
+export default BookPage;
+
+export { default as BookTable } from './BookTable';
