+++ frontend/src/components/MakerPage/RangeSlider.js
