+++ frontend/src/components/MakerPage/index.ts
@@ -1,0 +1,5 @@
+import MakerPage from './MakerPage';
+export default MakerPage;
+
+export { default as MakerForm } from './MakerForm';
+export { default as AutocompletePayments } from './AutocompletePayments';
